update 20201201, now println! is available

+++ include/src/asm/io.rs
@@ -1,17 +1,17 @@
 #[inline]
-pub unsafe fn outb(value: u8, port: u8) {
+pub unsafe fn outb(value: u8, port: u16) {
     llvm_asm!("outb %al,%dx"::"{al}" (value),"{dx}" (port));
 }
 
 #[inline]
-pub unsafe fn inb(port: u8) -> u8 {
+pub unsafe fn inb(port: u16) -> u8 {
     let _v: u8;
     llvm_asm!("inb %dx,%al":"={al}" (_v):"{dx}" (port));
     _v
 }
 
 #[inline]
-pub unsafe fn outb_p(value: usize, port: u8) {
+pub unsafe fn outb_p(value: u8, port: u16) {
     llvm_asm!(r#"outb %al,%dx
                  jmp 1f
                  1:jmp 1f
@@ -20,7 +20,7 @@ pub unsafe fn outb_p(value: usize, port: u8) {
 }
 
 #[inline]
-pub unsafe fn inb_p(port: u8) -> u8 {
+pub unsafe fn inb_p(port: u16) -> u8 {
     let _v: u8;
     llvm_asm!(r#"inb %dx,%al
                 jmp 1f
